feat: update translations

+++ frontend/src/lib/utils/schemas.ts
@@ -291,7 +291,7 @@ export const EvidenceSchema = z.object({
 
 export const GeneralSettingsSchema = z.object({
 	security_objective_scale: z.string(),
-	ebios_radar_max: z.number().int().default(6),
+	ebios_radar_max: z.number().default(6),
 	ebios_radar_green_zone_radius: z.number().default(0.2),
 	ebios_radar_yellow_zone_radius: z.number().default(0.9),
 	ebios_radar_red_zone_radius: z.number().default(2.5)
